requirements.txt
albumentations==1.3.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
imageio==2.15.0
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python-headless==4.8.0.76
Pillow==8.4.0
pkg_resources==0.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0.1
qudida==0.0.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.1.1
